Schema def fixes, other startup errors addressed

+++ alarm/src/main/java/org/opennms/horizon/alarmservice/api/AlarmRestService.java
@@ -52,26 +52,24 @@ public interface AlarmRestService {
     @RolesAllowed({"user", "admin"})
     Response getAlarms(@Context final SecurityContext securityContext, final UriInfo uriInfo);
 
-    // OnmsAlarm get(Long id);  v1??
-
 
     @PUT
     @Path("{id}/memo") // curl -X PUT http://localhost:8181/cxf/alarmservice/alarms/1/memo -d 'user=mark&body=not null'
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
     @RolesAllowed({"user", "admin"})
-    Response updateMemo(@Context final SecurityContext securityContext, @PathParam("id") final Integer alarmId, final MultivaluedMapImpl params);
+    Response updateMemo(@Context final SecurityContext securityContext, @PathParam("id") final Long alarmId, final MultivaluedMapImpl params);
 
 
     @PUT
     @Path("{id}/journal") // curl -X PUT http://localhost:8181/cxf/alarmservice/alarms/1/journal -d 'user=mark&body=not null'
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
     @RolesAllowed({"user", "admin"})
-    Response updateJournal(@Context final SecurityContext securityContext, @PathParam("id") final Integer alarmId, final MultivaluedMapImpl params);
+    Response updateJournal(@Context final SecurityContext securityContext, @PathParam("id") final Long alarmId, final MultivaluedMapImpl params);
 
     @DELETE
     @Path("{id}/memo") // curl -X DELETE http://localhost:8181/cxf/alarmservice/alarms/1/memo -d 'user=mark&body=not null'
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
     @RolesAllowed({"admin"})
-    Response removeMemo(@Context final SecurityContext securityContext, @PathParam("id") final Integer alarmId);
+    Response removeMemo(@Context final SecurityContext securityContext, @PathParam("id") final Long alarmId);
 
 }
